Extracted sign in class to a separate component
